requirements.txt
langchain>=0.1.0
langchain_community
langchain_mcp_adapters>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
regex>=2023.0.0
mcp-server>=0.1.0
aioconsole>=0.7.0
